Update map.alias
Fixed syntax issues you brought up as well as I never use trueDistance but messing around with it found a bug when using the new -distance feature and fixed it. Also the copy for -objects stated you can add multiple at a time but that is not the case, so removed that.